LayoutTests/fast/forms/form-associated-element-removal.html

<!DOCTYPE html>
<script src="../js/resources/js-test-pre.js"></script>
<script>
description("Removing all children of a form should disassociate any form controls in that subtree");
var form = document.createElement('form');
var input = document.createElement('input');
form.appendChild(input);
form.innerHTML = '';
shouldBeNull("input.form");
shouldBe("form.elements.length", "0");
</script>
<script src="../js/resources/js-test-post.js"></script>
