third_party/blink/web_tests/platform/generic/fast/forms/form-associated-element-removal-expected.txt

Removing all children of a form should disassociate any form controls in that subtree

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input.form is null
PASS form.elements.length is 0
PASS successfullyParsed is true

TEST COMPLETE
